+++ engines/ags/engine/script/system_imports.cpp
@@ -20,8 +20,6 @@
  *
  */
 
-//include <stdlib.h>
-//include <string.h>
 #include "ags/engine/script/system_imports.h"
 
 namespace AGS3 {
@@ -96,12 +94,13 @@ int SystemImports::get_index_of(const String &name) {
 
 	if (name.GetLength() > 3) {
 		size_t c = name.FindCharReverse('^');
-		if (c != -1 && (c == name.GetLength() - 2 || c == name.GetLength() - 3)) {
+		if (c != (size_t)-1 && (c == name.GetLength() - 2 || c == name.GetLength() - 3)) {
 			// Function with number of prametrs on the end
 			// attempt to find it without the param count
 			return get_index_of(name.Left(c));
 		}
 	}
+
 	return -1;
 }
